remove separate script tag to avoid volar error
Volar shows errors when an interface is declared.
Interface declarations should have the export keyword before
them according to the vue docs:
https://vuejs.org/api/sfc-script-setup.html#typescript-only-features

Since the TableXProps interface depends on the TableCol type,
I exported that type as well. I'm not sure if this is the best way
to fix this issue but I vastly prefer this over a separate <script> tag