refactor checkbox functionality
The previous code seemed quite confusing to me.
It felt like it was not harnessing the powers of reactive properties;
instead treating them like regular variables - resulting in messy code.
This newer code uses reactivity and computed properties,
retains all previous functionality, and looks cleaner as a bonus.

The naming was also not the best.
I changed function and variable names to be as semantic as possible.